Much ARC conversion for the Find framework.
There are crashes stemming from objc_ptr, which are quite confusing and may be hard to fix. The good news is ARC completely replaces objc_ptr. The conflict seems to be with OakAppKit; if this were converted to ARC at the same time as Find, and objc_ptr removed entirely, it might work a lot better.