Enable responsive scrolling
When scrolling “bounces”, the overdraw area is not drawn by the document view; it simply shows whatever is underneath. For that reason, rendering of the document’s background color was moved to the document controller / window.

This is achieved with a custom contentView which renders the standard window background, but uses the document’s background color when drawing underneath the layout view. The scroll view itself is transparent.